% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_points.R
\name{clean_points}
\alias{clean_points}
\title{Clean occurrence records}
\usage{
clean_points(coord, merge_dist = 2, reference_layer, layer_filter = NULL,
  dist_unit = "km")
}
\arguments{
\item{coord}{data.frame. Data frame containing longitudes (Lon) and latitudes (Lat) of occurrence records of a species.}

\item{merge_dist}{numeric. Maximum distance between points to be merged.}

\item{reference_layer}{Raster* object. Any raster for the area containing the occurrence records.}

\item{layer_filter}{Values for locations at reference raster which occurrence records should be eliminated.}

\item{dist_unit}{character. Unit for the merging distance. Either 'km' for kilometers or 'degrees' for decimal degrees.}
}
\value{
Data frame with remaining longitudes and latitudes.
}
\description{
\code{clean_points} Eliminates species occurrence records that are too close to each other or at undesired locations.
}
\examples{

FulanusEcoRasters_present <-
  get_rasters(
    raster_source = "/Users/gabriel/Documents/Mapinguari/global_grids_10_minutes",
    ext = FulanusDistribution,
    margin = 5,
    non_fixed_var = c('prec', 'tmin', 'tmax'),
    fixed_var = 'alt',
    years = c("present"),
    reorder = TRUE)

alt <- FulanusEcoRasters_present$present$alt

clean_points(coord = FulanusDistribution,
  merge_dist = 2,
  reference_layer = alt > 500 & alt < 1000,
  layer_filter = 0)

}
