% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardFUNs.R
\name{daylengthFUN}
\alias{daylengthFUN}
\title{Gets average month day lengths for an area}
\usage{
daylengthFUN(reference_raster)
}
\arguments{
\item{reference_raster}{RasterStack or RasterLayer. Any raster representative of the area you want day lengths to. The day length raster will have the same projection, resolution and extension.}
}
\value{
Returns a RasterStack with 12 layers, one for each month, containing information on the duration of the day at each pixel.
}
\description{
\code{daylengthFUN} Generates surfaces with information on day length for each month accross an area.
}
\examples{
daylengthFUN(df_alt)

}
