% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_alt.R
\docType{data}
\name{df_alt}
\alias{df_alt}
\title{Altitude for Distrito Federal, Brazil}
\format{A raster layer with 13 rows and 24 columns}
\usage{
df_alt
}
\description{
CGIAR-SRTM Altitude data (Jarvis et al, 2010) resampled to 2.5 minute resolution for the Distrito Federal region in central Brazil from 1960 to 1990.
}
\references{
Jarvis et al (2005) <http://srtm.csi.cgiar.org.>
}
\keyword{datasets}
