% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_tmin.R
\docType{data}
\name{df_tmin}
\alias{df_tmin}
\title{Minimum temperature for Distrito Federal, Brazil}
\format{A raster stack with 12 layers, 13 rows and 24 columns}
\usage{
df_tmin
}
\description{
Average monthly minimum temperature data from WorldClim 1.4 (Hijmans et al, 2005, licensed under CC BY-SA 4.0) in 2.5 minute resolution for the Distrito Federal region in central Brazil from 1960 to 1990.
}
\references{
Hijmans et al (2005) <doi.org/10.1002/joc.1276>
}
\keyword{datasets}
