% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoPhysiology.R
\name{EcoPhysiology}
\alias{EcoPhysiology}
\title{Transform environmental rasters in ecophysiological rasters.}
\usage{
EcoPhysiology(raster_source, Perf_args, separator, PerfFUN)
}
\arguments{
\item{raster_source}{character or list of RasterStack. You can either input a path to a folder with the required rasters or a list of RasterStack organized by year/scenario.}

\item{separator}{character. Character that separates variable names, years and scenarios.}

\item{PerfFUN}{}
}
\value{
Returns a list of raster stacks for the variables required, organized by year/scenario combination.
}
\description{
\code{EcoPhysiology} Transform environmental rasters in ecophysiological rasters.
}
\examples{
FulanusEcoRasters <-
  Ecology(raster_source = "/Users/gabriel/Documents/Mapinguari-development/global_grids_10_minutes",
    ext = FulanusDistribution,
    non_fixed_var = c('prec', 'tmin', 'tmax'),
    fixed_var = 'alt',
    years = c("present", '2050', '2070'),
    scenarios = c('rcp26', 'rcp45', 'rcp85'),
    phenology = 'month',
    reorder = TRUE)

Perf_rasters <-
EcoPhysiology(raster_source = FulanusEcoRasters,
 Perf_args = list(temp = 'tmax', size = 10),
 separator = '_',
 PerfFUN = perf_functions$predict$model_1)

}
