% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curves.R
\name{fit_curves}
\alias{fit_curves}
\title{Creates functions relating environment to physiology.}
\usage{
fit_curves(formula, data, type, fixed = NULL, random = NULL, start = NULL,
  correlation = NULL, ...)
}
\arguments{
\item{formula}{formula. Model formula.}

\item{data}{data frame. Table containing variables in model formula.}

\item{type}{character. Algorithm used for model. 'GAMM' or 'NLME'}

\item{fixed}{named list. See \code{nlme::nlme}}

\item{random}{named list. See \code{mgcv::gamm} and \code{nlme::nlme}}

\item{start}{named list. See \code{nlme::nlme}}

\item{correlation}{named list. See \code{mgcv::gamm} and \code{nlme::nlme}}

\item{...}{further parameters for \code{mgcv::gamm} or \code{nlme::nlme}}
}
\value{
Returns a tibble containing model specifications, statistics and a predictor function.
}
\description{
\code{fit_curves} Creates functions relating environment to physiology. You can run multiple models if you input their arguments as lists.
}
\examples{
perf_functions <-
  fit_curves(formula = performance ~ s(temp, bs = 'cs') + size,
    data = FulanusPhysiology,
    type = 'GAMM',
    random = list(id = ~ 1)
  )

perf_functions <-
  fit_curves(formula = performance ~ s(temp, bs = 'cs') + size,
    data = FulanusPhysiology,
    type = 'GAMM',
    random = list(id = ~ 1),
    correlation = list(nlme::corAR1(form = ~ 1 | id))
  )

formula_list <-
  list(
    performance ~ s(temp, bs = 'cs'),
    performance ~ s(temp, bs = 'cs') + size,
    performance ~ s(temp, bs = 'cs') + hydration,
    performance ~ s(temp, bs = 'cs') + hydration + size
    )

perf_functions <-
  fit_curves(formula = formula_list,
    data = FulanusPhysiology,
    type = 'GAMM',
    random = list(id = ~ 1),
    correlation = list(nlme::corAR1(form = ~ 1 | id)
  ))

correlation_list <- list(a = nlme::corAR1(form = ~ 1 | id),
  b = nlme::corAR1(0.1, form = ~ 1 | id),
  c = nlme::corARMA(form = ~ 1 | id),
  d = nlme::corARMA(0.1, form = ~ 1 | id))

perf_functions <-
  fit_curves(formula = formula_list,
    data = FulanusPhysiology,
    type = 'GAMM',
    random = list(id = ~ 1),
    correlation = correlation_list
  )

}
