% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FulanusAgar.R
\docType{data}
\name{FulanusAgar}
\alias{FulanusAgar}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
FulanusAgar
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
