% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FulanusDistribution.R
\docType{data}
\name{FulanusDistribution}
\alias{FulanusDistribution}
\title{615 occurrence records}
\format{A data frame with 615 rows and 3 variables:
\describe{
  \item{species}{species name}
  \item{Lon}{Longitude of occurrence point}
  \item{Lat}{Latitude of occurrence point}
  ...
}}
\usage{
FulanusDistribution
}
\description{
A dataset containing simulated distribution records
}
\keyword{datasets}
