% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FulanusMicroclimate.R
\docType{data}
\name{FulanusMicroclimate}
\alias{FulanusMicroclimate}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 9360 rows and 16 variables:
\describe{
  \item{species}{price, in US dollars}
  \item{day}{weight of the diamond, in carats}
  \item{species}{price, in US dollars}
  \item{day}{weight of the diamond, in carats}
  \item{species}{price, in US dollars}
  \item{day}{weight of the diamond, in carats}
  \item{species}{price, in US dollars}
  \item{day}{weight of the diamond, in carats}
  \item{species}{price, in US dollars}
  \item{day}{weight of the diamond, in carats}
  \item{species}{price, in US dollars}
  \item{day}{weight of the diamond, in carats}
  \item{species}{price, in US dollars}
  \item{day}{weight of the diamond, in carats}
  \item{species}{price, in US dollars}
  \item{day}{weight of the diamond, in carats}
  ...
}}
\usage{
FulanusMicroclimate
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
