% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FulanusPhysiology.R
\docType{data}
\name{FulanusPhysiology}
\alias{FulanusPhysiology}
\title{Running speed achieved by 125 *Fulanus beltranus* lizards in 681 trials under different temperatures.}
\format{A data frame with 681 rows and 8 variables:
\describe{
  \item{species}{price, in US dollars}
  \item{id}{weight of the diamond, in carats}
  \item{temp}{price, in US dollars}
  \item{performance}{weight of the diamond, in carats}
  \item{size}{weight of the diamond, in carats}
  \item{site}{weight of the diamond, in carats}
  \item{acctemp}{weight of the diamond, in carats}
  \item{hydration}{hydration station during trial, in carats}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
FulanusPhysiology
}
\description{
A dataset containing the running speed of *Fulanus beltranus* lizards under different temperatures, the temperatures of the runs, individual identities for each lizard, body size of each individual, the tempetature they were acclimated to, the site where they were captured, their hydration state during the trial.
}
\keyword{datasets}
