% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapinguari.R
\docType{package}
\name{Mapinguari}
\alias{Mapinguari}
\alias{Mapinguari-package}
\title{Mapinguari: Tools for process-based biogeographical analysis.}
\description{
Mapinguari provides solutions for incorporating biological processes in biogeographical analysis.
}
\section{Mapinguari functions}{


\itemize{
  \item \code{\link{clean_points}}
  \item \code{\link{fit_curves}}
  \item \code{\link{get_rasters}}
  \item \code{\link{pseudoabsences}}
  \item \code{\link{qlist}}
  \item \code{\link{transform_rasters}}
  \item \code{\link{daylengthFUN}}
  \item \code{\link{AETFUN}}
  \item \code{\link{sradFUN}}
  \item \code{\link{rhFUN}}
  }
}

