% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardFUNs.R
\name{PC2FMFUN}
\alias{PC2FMFUN}
\title{Physical Chemistry Fire Frequency Model (PC2FM) by Guyette (2012)}
\usage{
PC2FMFUN(prec, temp, alt)
}
\arguments{
\item{prec}{numeric. Precipitation raster.}

\item{temp}{numeric. Temperature raster.}

\item{alt}{numeric. Altitude raster.}
}
\value{
numeric. Fire frequency based on physical chemical factors
}
\description{
\code{PC2FMFUN} Applies Guyette's Fire frequency model (Guyette, Richard P., et al. "Predicting fire frequency with chemistry and climate." Ecosystems 15.2 (2012): 322-335.)
}
\examples{
PC2FMFUN(50, 25, 1000)

}
