% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardFUNs.R
\name{PETFUN}
\alias{PETFUN}
\title{Spatialize \code{PET_fromTemp} function from package \code{EcoHydrology}}
\usage{
PETFUN(tmax, tmin, alt, separator = "_")
}
\arguments{
\item{tmax}{Raster* object. Maximum temperature raster.}

\item{tmin}{Raster* object. Minimum temperature raster.}

\item{alt}{Raster* object. Altitude raster.}

\item{separator}{character. Character that separates variable names, and scenarios.}
}
\value{
Returns a RasterLayer with estimates of Potential EvapoTranspiration in milimiters.
}
\description{
\code{PETFUN} Gets Potential EvapoTranspiration (PET) rasters from maximum temperature, minimum temperature and altitude rasters by applying function \code{PET_fromTemp} from package \code{EcoHydrology}
}
\examples{
PET <- PETFUN(df_tmax, df_tmin, df_alt)

}
