% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoabsences.R
\name{pseudoabsences}
\alias{pseudoabsences}
\title{Generate background pseudoabsence records}
\usage{
pseudoabsences(coord, diameter, n)
}
\arguments{
\item{coord}{Coordinates of points to generate pseudoabsence points around.}

\item{diameter}{Diameter of circles around coordinate points.}

\item{n}{Number of pseudoabsences to be generated.}
}
\value{
Data frame with coordinates of pseudoabsence points
}
\description{
\code{pseudoabsences} Creates random pseudoabsence coordinates within a radial distance from specified points.
}
\examples{

pseudoabsences(FulanusDistribution[2:3], 150, 2000)

}
