% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardFUNs.R
\name{sin_h}
\alias{sin_h}
\title{Sinervo (2010) hours of activity model}
\usage{
sin_h(tmax, tmin, thrs, res)
}
\arguments{
\item{tmax}{Raster* object. Maximum temperature raster.}

\item{tmin}{Raster* object. Minimum temperature raster.}

\item{thrs}{numeric. Temperature threshold in same unit as rasters.}

\item{res}{numeric. time resolution in parts of hour.}
}
\value{
numeric. Amount of time in hours above temperature threshold in simulated daily temperature variation.
}
\description{
\code{sin_h} Simulates daily variation in temperature and counts amount of time above a temperature threshold, as seen in Sinervo, Barry, et al. "Erosion of lizard diversity by climate change and altered thermal niches." Science 328.5980 (2010): 894-899.
}
\examples{
sin_h(28, 10, 23, 3)

}
