% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardFUNs.R
\name{sradFUN}
\alias{sradFUN}
\title{Spatialize \code{Solar} function from package \code{EcoHydrology}}
\usage{
sradFUN(alt, tmax, tmin, separator = "_")
}
\arguments{
\item{alt}{Raster* object. Altitude raster.}

\item{tmax}{Raster* object. Maximum temperature raster.}

\item{tmin}{Raster* object. Minimum temperature raster.}

\item{separator}{character. Character that separates variable names, and scenarios.}
}
\value{
Returns a RasterLayer with estimates of Solar in kiloJoules by square meter by day.
}
\description{
\code{sradFUN} Gets estimates of solar radiation for an area, based on maximum and minimum temperatures and altitude rasters by applying function \code{Solar} from package \code{EcoHydrology}
}
\examples{
srad <- sradFUN(df_alt, df_tmax, df_tmin)

}
