% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TtorquatusBreeding.R
\docType{data}
\name{TtorquatusBreeding}
\alias{TtorquatusBreeding}
\title{Breeding status of *Tropidurus torquatus* lizards at 15 locations in Brazil during each month of the year.}
\format{A data frame with 15 rows and 14 variables:
\describe{
  \item{Lon}{Longitude of occurrence records in decimal degrees}
  \item{Lat}{Latitude of occurrence records in decimal degrees}
  \item{January}{Binary breeding status at each location for the month of January, 1 means breeding, 0 means not breeding}
  \item{February}{Binary breeding status at each location for the month of February, 1 means breeding, 0 means not breeding}
  \item{March}{Binary breeding status at each location for the month of March, 1 means breeding, 0 means not breeding}
  \item{April}{Binary breeding status at each location for the month of April, 1 means breeding, 0 means not breeding}
  \item{May}{Binary breeding status at each location for the month of May, 1 means breeding, 0 means not breeding}
  \item{June}{Binary breeding status at each location for the month of June, 1 means breeding, 0 means not breeding}
  \item{July}{Binary breeding status at each location for the month of July, 1 means breeding, 0 means not breeding}
  \item{August}{Binary breeding status at each location for the month of August, 1 means breeding, 0 means not breeding}
  \item{September}{Binary breeding status at each location for the month of September, 1 means breeding, 0 means not breeding}
  \item{October}{Binary breeding status at each location for the month of October, 1 means breeding, 0 means not breeding}
  \item{November}{Binary breeding status at each location for the month of November, 1 means breeding, 0 means not breeding}
  \item{December}{Binary breeding status at each location for the month of December, 1 means breeding, 0 means not breeding}
  ...
}}
\usage{
TtorquatusBreeding
}
\description{
A dataset containing information on if *Tropidurus torquatus* is breeding or not at specific locations and times.
}
\keyword{datasets}
