% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TtorquatusGradient.R
\docType{data}
\name{TtorquatusGradient}
\alias{TtorquatusGradient}
\title{*Tropidurus torquatus* body temperatures at temperature gradient experiments.}
\format{
A data frame with 3443 rows and 3 variables:
\describe{
  \item{id}{individual identity of the lizard perform}
  \item{temp}{lizard body temperature at the moment of the trial, in Celsius}
  \item{site}{place where lizard was collected}
  ...
}
}
\usage{
TtorquatusGradient
}
\description{
A dataset containing 3443 body temperature records of 52 *Tropidurus torquatus*  from 6 localities at temperature gradients.
}
\keyword{datasets}
