% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TtorquatusOperative.R
\docType{data}
\name{TtorquatusOperative}
\alias{TtorquatusOperative}
\title{Operative temperatures of multiple microhabitats at 6 localities in Brazil from 2014 to 2016.}
\format{A data frame with 915684 rows and 13 variables:
\describe{
  \item{site}{place where temperatures were collected}
  \item{description}{description of site}
  \item{Lon}{Longitude of sampling point}
  \item{Lat}{Latitude of sampling point}
  \item{temp}{temperature at microhabitat, in degrees Celsius}
  \item{microhabitat}{microhabitat sampled}
  \item{year}{year of sampling}
  \item{month}{month of sampling}
  \item{day}{day of sampling}
  \item{hour}{hour of sampling}
  \item{minute}{minute of sampling}
  \item{t_air_max}{maximum air temperature of the day at nearest weather station, in degrees Celsius}
  ...
}}
\usage{
TtorquatusOperative
}
\description{
A dataset containing operative temperatures of multiple microhabitats at 6 localities in Brazil from 2014 to 2016.
}
\keyword{datasets}
