% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TtorquatusPerformance.R
\docType{data}
\name{TtorquatusPerformance}
\alias{TtorquatusPerformance}
\title{Running speed achieved by 72 *Tropidurus torquatus* lizards in 274 trials under different temperatures.}
\format{A data frame with 274 rows and 6 variables:
\describe{
  \item{species}{species name}
  \item{id}{individual identity of the lizard perform}
  \item{temp}{lizard body temperature at the moment of the trial, in Celsius}
  \item{performance}{maximum running speed at trial, in meters per second}
  \item{size}{lizard body size, in centimeters}
  \item{site}{place where lizard was collected}
  ...
}}
\usage{
TtorquatusPerformance
}
\description{
A dataset containing 274 running speed trials of *Tropidurus torquatus* lizards under different temperatures, the temperatures of the runs, individual identities for each lizard, body size of each individual and the site where they were captured.
}
\keyword{datasets}
