% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardFUNs.R
\name{AETFUN}
\alias{AETFUN}
\title{Generates Actual EvapoTranspiration rasters}
\usage{
AETFUN(PET, prec)
}
\arguments{
\item{PET}{RasterStack with 12 layers. Total month Potential EvapoTranspiration rasters.}

\item{prec}{RasterStack with 12 layers. Total month precipitation rasters.}
}
\value{
Returns a RasterLayer with estimates of Actual EvapoTranspiration in milimiters.
}
\description{
\code{AETFUN} Applies Duncan Golicher's Bucket model to Potential EvapoTranspiration and precipitation rasters in order to get Actual Evapotranspiration estimates for an area.
}
\examples{

FulanusEcoRasters_present <-
  get_rasters(
    raster_source = "/Users/gabriel/Documents/Mapinguari/global_grids_10_minutes",
    ext = FulanusDistribution,
    margin = 5,
    non_fixed_var = c('prec', 'tmin', 'tmax'),
    fixed_var = 'alt',
    years = c("present"),
    reorder = TRUE)

alt <- FulanusEcoRasters_present$present$alt
tmax <- FulanusEcoRasters_present$present[[25:36]]/10
tmin <- FulanusEcoRasters_present$present[[13:24]]/10
prec <- FulanusEcoRasters_present$present[[1:12]]

PET <- PETFUN(tmax, tmin, alt)

AETFUN(PET, prec)

AETFUN(PET, prec)

}
