% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardFUNs.R
\name{daylengthFUN}
\alias{daylengthFUN}
\title{Gets average month day lengths for an area}
\usage{
daylengthFUN(reference_raster)
}
\arguments{
\item{reference_raster}{RasterStack or RasterLayer. Any raster representative of the are you want daylengths to.}
}
\value{
Returns a RasterStack with 12 layers, one for each month, containing information on the duration of the day at each pixel.
}
\description{
\code{daylengthFUN} Generates surfaces with information on day length for each month accross an area.
}
\examples{
\dontrun{
FulanusEcoRasters_present <-
  get_rasters(
    var = c('alt', 'prec', 'tmax', 'tmin'),
    scenarios = c("present"),
    source = "C:/Users/gabri/Dropbox/Mapinguari/global_grids_10_minutes",
    ext = FulanusDistribution[c(2,3)],
    margin = 5,
    reorder = c(1, 10, 11, 12, 2, 3, 4, 5, 6, 7, 8, 9))

daylengthFUN(FulanusEcoRasters_present$present)
}
}
