% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardFUNs.R
\name{daylengthFUN}
\alias{daylengthFUN}
\title{Gets average month day lengths for an area}
\usage{
daylengthFUN(reference_raster)
}
\arguments{
\item{reference_raster}{RasterStack or RasterLayer. Any raster representative of the are you want daylengths to.}
}
\value{
Returns a RasterStack with 12 layers, one for each month, containing information on the duration of the day at each pixel.
}
\description{
\code{daylengthFUN} Generates surfaces with information on day length for each month accross an area.
}
\examples{

FulanusEcoRasters_present <-
  get_rasters(
    raster_source = "/Users/gabriel/Documents/Mapinguari/global_grids_10_minutes",
    ext = FulanusDistribution,
    margin = 5,
    non_fixed_var = c('prec', 'tmin', 'tmax'),
    fixed_var = 'alt',
    years = c("present"),
    reorder = TRUE)

daylengthFUN(FulanusEcoRasters_present$present)

}
