% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curves.R
\name{fit_curves}
\alias{fit_curves}
\title{Creates vectorized predict functions from models.}
\usage{
fit_curves(models, predict_formals = NULL, separator = "_")
}
\arguments{
\item{models}{list. List with models that have methods for function `predict`.}

\item{predict_formals}{character. Names of variables necessary for predict function. If NULL, function will attempt to retrieve them from the model's formula.}

\item{separator}{character. Character that separates variable names, years and scenarios.}
}
\value{
Returns a list with a table of statistics for model comparision, and a list for each model containing the original call, its output and a vectorized function that gets predictions for the model.
}
\description{
\code{fit_curves} Takes inputed models and create vectorized functions able to get the model predictionsfor any value inputed. Also outputs a table comparing models.
}
\examples{

library(mgcv)

  fit_curves(FUN_qlist = qlist(tpc_size = gamm(performance ~ s(temp, bs = 'cs') + size,
                                               random = list(id = ~ 1),
                                               data = FulanusPhysiology),
                               tpc_no_size = gamm(performance ~ s(temp, bs = 'cs'),
                                                  random = list(id = ~ 1)),
                                                  data = FulanusPhysiology))

}
