% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rasters.R
\name{get_rasters}
\alias{get_rasters}
\title{Retrieve and organize spatial environmental information.}
\usage{
get_rasters(raster_source = NULL, ext = c(-180, 180, -60, 90), margin = 0,
  resolution = 10, non_fixed_var = NULL, fixed_var = NULL,
  years = "present", scenarios = NULL, baseline = "present",
  separator = "_", projection_model = "MP", seasons = NULL,
  reorder = FALSE, seasons_args = NULL, summary_args = NULL,
  summaryFUN = NULL, download_source = "pkg-raster", alert = NULL)
}
\arguments{
\item{raster_source}{character or list of RasterStack. You can either input a path to a folder with the required rasters or a list of RasterStack organized by year/scenario.}

\item{ext}{numeric vector or dataframe with coordinates. Extension to crop rasters. Input can be either a vector with limiting latitudes and longitudes or a table of coordinates. In case the input is a numeric vector, the order of values should be: western most longitude, easter most longitude, southern most latitude then northern most latitude. Longitude values west of GMT and latitude values south of equator are distinguished by a negative sign. In case the input is a table of coordinates, the points further in each direction will determine the extent for cropping.}

\item{margin}{numeric. Additional distance to be added to margin of extent, in degrees.}

\item{resolution}{numeric. Spatial resolution of rasters, in degrees. This is used for downloading only.}

\item{non_fixed_var}{character vector. Names of the time varying variables to be outputed. This variables will be subject to variation according to argument `years`.}

\item{fixed_var}{character vector. Names of the time fixed variables to be outputed. This variables will not be subject to variation according to argument `years`.}

\item{years}{character vector. Names of years for time varying variables.}

\item{scenarios}{character vector. Names of future scenarios to for time varying variables.}

\item{baseline}{character vector. Names of years not subject to scenarios.}

\item{separator}{character. Character that separates variable names, years and scenarios.}

\item{projection_model}{character. Projection model for future variables. Must be any model accepted in \code{raster::getData}}

\item{seasons}{numerical vector or function. Months at the beggining and end of phenological events, or function relating phenological event to environmental conditions. If NULL, no summarizing is made (remains by month).}

\item{reorder}{logical. If TRUE, will use last two characters of layer names in RasterStacks with 12 layers to order them in ascending order.}

\item{seasons_args}{named list of strings. Correspondence between PhenFUN arguments and raster names.}

\item{summary_args}{named list. Additional arguments for summarizing function.}

\item{summaryFUN}{function. Function used to summarize months inside seasons.}

\item{download_source}{character. Where to download rasters from. Default is 'pkg-raster' which will get rasters from WorldClim, using package 'raster' functions.}

\item{alert}{integer. Plays a sound alert when function is done running. See documentation of package beepr for description of sound options.}

\item{reorder}{logical. If TRUE, reorders layers by last two characters in their names, after replacing letters for 0.}

\item{alert}{integer. If an integer is input, the functino will play a sound after it is finished. See package 'beepr' for correspondence between numbers and sounds.}
}
\value{
Returns a list of raster stacks for the variables required, organized by year/scenario combination.
}
\description{
\code{get_rasters} returns an organized list of environmental rasters.
}
\examples{

Fulanus_Ecorasters_download <-
  get_rasters(
    ext = FulanusDistribution,
    margin = 5,
    non_fixed_var = c('prec', 'tmin', 'tmax'),
    years = c("present", '2050', '2070'),
    scenarios = c('rcp26', 'rcp45', 'rcp85'),
    alert = 6)

Fulanus_Ecorasters_dir <-
  get_rasters(
    raster_source = "/Users/gabriel/Documents/Mapinguari/global_grids_10_minutes",
    ext = FulanusDistribution,
    margin = 5,
    non_fixed_var = c('prec', 'tmin', 'tmax'),
    fixed_var = "alt",
    years = c("present", '2050', '2070'),
    scenarios = c('rcp26', 'rcp45', 'rcp85'),
    reorder = TRUE,
    alert = 4)


Fulanus_EcoRasters_Phenology_mean_sd <-
  get_rasters(raster_source = Fulanus_Ecorasters_dir,
    ext = FulanusDistribution,
    margin = 5,
    non_fixed_var = c('prec', 'tmin', 'tmax'),
    years = c("present", '2050', '2070'),
    scenarios = c('rcp26', 'rcp45', 'rcp85'),
    seasons = list(breeding = c(3:8), non_breeding = c(9:12, 1, 2)),
    summaryFUN = list(tmax = c("mean", "sd"), tmin = c("mean", "sd"), prec = "sum"))


}
