% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardFUNs.R
\name{PC2FMFUN}
\alias{PC2FMFUN}
\title{Guyette's Physical Chemistry Fire Frequency Model (PC2FM)}
\usage{
PC2FMFUN(prec, temp, alt)
}
\arguments{
\item{prec}{numeric. Precipitation raster.}

\item{temp}{numeric. Temperature raster.}

\item{alt}{numeric. Altitude raster.}
}
\value{
numeric. Fire frequency based on physical chemical factors
}
\description{
\code{PC2FMFUN} Applies Guyette's Fire frequency model (Guyette et al, 2012).
}
\examples{
PC2FMFUN(50, 25, 1000)

}
