% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardFUNs.R
\name{PETFUN}
\alias{PETFUN}
\title{Gets average month daylengths for an area}
\usage{
PETFUN(tmax, tmin, alt, forest = 0)
}
\arguments{
\item{tmax}{RasterStack or RasterLayer. Any raster representative of the are you want daylengths to.}

\item{tmin}{RasterStack or RasterLayer. Any raster representative of the are you want daylengths to.}

\item{alt}{RasterStack or RasterLayer. Any raster representative of the are you want daylengths to.}
}
\value{
Returns a RasterStack with 12 layers, one for each month, containing information on the duration of the day at each pixel.
}
\description{
\code{PETFUN} Get model predictions for a raster stack
}
\examples{

FulanusEcoRasters_present <-
  get_rasters(
    raster_source = "/Users/gabriel/Documents/Mapinguari/global_grids_10_minutes",
    ext = FulanusDistribution,
    margin = 5,
    non_fixed_var = c('prec', 'tmin', 'tmax'),
    fixed_var = 'alt',
    years = c("present"),
    reorder = TRUE)

alt <- FulanusEcoRasters_present$present$alt
tmax <- FulanusEcoRasters_present$present$tmax_1/10
tmin <- FulanusEcoRasters_present$present$tmin_1/10

PETFUN(tmax, tmin, alt)

}
