% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardFUNs.R
\name{PETFUN}
\alias{PETFUN}
\title{Wrapper around \code{PET_fromTemp} function from package \code{EcoHydrology}}
\usage{
PETFUN(tmax, tmin, alt, separator = "_")
}
\arguments{
\item{tmax}{Raster* object. Maximum temperature raster.}

\item{tmin}{Raster* object. Minimum temperature raster.}

\item{alt}{Raster* object. Altitude raster.}

\item{separator}{character. Character that separates variable names, and scenarios.}
}
\value{
Returns a RasterLayer with estimates of Potential EvapoTranspiration in milimiters.
}
\description{
\code{PETFUN} Gets Potential EvapoTranspiration (PET) rasters from maximum temperature, minimum temperature and altitude rasters by applying function \code{PET_fromTemp} from package \code{EcoHydrology}
}
\examples{
\dontrun{
FulanusEcoRasters_present <-
  get_rasters(
    var = c('alt', 'prec', 'tmax', 'tmin'),
    scenarios = c("present"),
    source = "C:/Users/gabri/Dropbox/Mapinguari/global_grids_10_minutes",
    ext = FulanusDistribution[c(2,3)],
    margin = 5,
    reorder = c(1, 10, 11, 12, 2, 3, 4, 5, 6, 7, 8, 9))

alt <- FulanusEcoRasters_present$present$alt
tmin <- FulanusEcoRasters_present$present[[25:36]]/10
tmax <- FulanusEcoRasters_present$present[[13:24]]/10

PET <- PETFUN(tmax, tmin, alt)
}
}
