% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenology.R
\name{Phenology}
\alias{Phenology}
\title{Groups rasters by duration of phenological event.}
\usage{
Phenology(raster_list, PhenFUN, Phen_args)
}
\arguments{
\item{raster_list}{list of RasterStacks. Only accept stacks with 12 layers.}

\item{PhenFUN}{function. Function relating phenological event to environmental conditions.}

\item{Phen_args}{named list of strings. Correspondence between PhenFUN arguments and raster names.}
}
\value{
Returns a list of raster stacks for the variables required, organized by year/scenario combination.
}
\description{
\code{Phenology} Groups rasters by duration of phenological event.
}
\examples{

}
