% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{Physiology}
\alias{Physiology}
\title{Creates functions relating environment to physiology.}
\usage{
Physiology(formula, data, type, fixed = NULL, random = NULL, start = NULL,
  correlation = NULL, ...)
}
\arguments{
\item{formula}{formula. Model formula.}

\item{data}{data frame. Table containing variables in model formula.}

\item{type}{character. Algorithm used for model. 'GAMM' or 'NLME'}

\item{fixed}{named list. See \code{mgcv::gamm}}

\item{random}{named list. See \code{mgcv::gamm} and \code{nlme::nlme}}

\item{start}{named list. See \code{nlme::nlme}}

\item{correlation}{named list. See \code{mgcv::gamm} and \code{nlme::nlme}}

\item{...}{further parameters for \code{mgcv::gamm} or \code{nlme::nlme}}
}
\value{
Returns a list of raster stacks for the variables required, organized by year/scenario combination.
}
\description{
\code{Physiology} Creates functions relating environment to physiology.
}
\examples{
perf_functions <-
Physiology(formula = performance ~ s(temp, bs = 'cs') + size,
 data = FulanusPhysiology,
 type = 'GAMM',
 random = list(id = ~ 1)
)

}
