% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{Physiology}
\alias{Physiology}
\title{Creates functions relating environment to physiology.}
\usage{
Physiology(formula, data, type, fixed = NULL, random = NULL, start = NULL,
  correlation = NULL, x_axis = NULL, y_axis = NULL, z_axis = NULL,
  scale = NULL, ...)
}
\arguments{
\item{...}{}
}
\value{
Returns a list of raster stacks for the variables required, organized by year/scenario combination.
}
\description{
\code{Physiology} Creates functions relating environment to physiology.
}
\examples{

perf_functions <-
Physiology(formula = performance ~ s(temp, bs = 'cs') + size,
 data = FulanusPhysiology,
 type = 'GAMM',
 random = list(id = ~ 1)
)

}
