% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardFUNs.R
\name{sradFUN}
\alias{sradFUN}
\title{Wrapper around \code{Solar} function from package \code{EcoHydrology}}
\usage{
sradFUN(alt, tmax, tmin)
}
\arguments{
\item{alt}{Raster* object. Altitude raster.}

\item{tmax}{Raster* object. Maximum temperature raster.}

\item{tmin}{Raster* object. Minimum temperature raster.}
}
\value{
Returns a RasterLayer with estimates of Solar in kiloJoules by square meter by day.
}
\description{
\code{sradFUN} Gets estimates of solar radiation for an area, based on maximum and minimum temperatures and altitude rasters by applying function \code{Solar} from package \code{EcoHydrology}
}
\examples{
FulanusEcoRasters_present <-
  get_rasters(
    raster_source = "/Users/gabriel/Documents/Mapinguari/global_grids_10_minutes",
    ext = FulanusDistribution,
    margin = 5,
    non_fixed_var = c('prec', 'tmin', 'tmax'),
    fixed_var = 'alt',
    years = c("present"),
    reorder = TRUE)

alt <- FulanusEcoRasters_present$present$alt
tmax <- FulanusEcoRasters_present$present[[25:36]]/10
tmin <- FulanusEcoRasters_present$present[[13:24]]/10

srad <- sradFUN(alt = alt, tmax = tmax, tmin = tmin)

}
