% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rasters.R
\name{summarize_rasters}
\alias{summarize_rasters}
\title{Groups rasters by duration of phenological event.}
\usage{
summarize_rasters(raster_stack, seasons = list(year = c(1:12)),
  seasons_args = NULL, summaryFUN = c(mean), summary_args = NULL,
  separator = "_", time_res = 12)
}
\arguments{
\item{raster_stack}{RasterStack.}

\item{seasons}{numerical vector or function. Months at the beggining and end of phenological events, or function relating phenological event to environmental conditions. If NULL, no summarizing is made (remains by month).}

\item{summaryFUN}{function. Function used to summarize months inside seasons.}

\item{summary_args}{named list. Additional arguments for summarizing function.}

\item{separator}{character. Character that separates variable names, years and scenarios.}

\item{time_res}{integer. How many layers do time varying variables have? Default is 12, as in 12 months in a year, the time resolution in WorldClim.}

\item{season_args}{named list. Correspondence between 'seasons' arguments and raster names.}
}
\value{
Returns a list of raster stacks for the variables required, organized by year/scenario combination.
}
\description{
\code{summarize_rasters} Summarize rasters by duration of phenological event.
}
\examples{

Fulanus_Ecorasters_download <-
  get_rasters(
    ext = FulanusDistribution,
    margin = 5,
    non_fixed_var = c('prec', 'tmin', 'tmax'),
    years = c("present", '2050', '2070'),
    scenarios = c('rcp26', 'rcp45', 'rcp85'),
    alert = 6)

Phenology_mean_sd <-
  lapply(Fulanus_Ecorasters_download, summarize_rasters,
    seasons = list(breeding = c(3:8), non_breeding = c(9:12, 1, 2)),
    summaryFUN = list(tmax = c("mean", "sd"), tmin = c("mean", "sd"), prec = "sum"))

Phenology_weighted_mean <-
  lapply(Fulanus_Ecorasters_download, summarize_rasters,
    seasons = list(breeding = c(3:8), non_breeding = c(9:12, 1, 2)),
    summaryFUN = "weighted.mean",
    summary_args = list(w = c(0.5, rep(1, 4), 0.5))
  )

PhenFUN <- function(x) 1/(1 + exp((150 - x)/2))

Phenology_by_precFUN <-
  lapply(Fulanus_Ecorasters_download, summarize_rasters,
    seasons = list(rainy_season = PhenFUN),
    seasons_args = list(x = 'prec'),
    summaryFUN = "raster::weighted.mean",
    summary_args = list(w = "rainy_season")
  )

}
