% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_rasters.R
\name{transform_rasters}
\alias{transform_rasters}
\title{Transform environmental rasters in ecophysiological rasters.}
\usage{
transform_rasters(raster_source, Perf_args, separator, PerfFUN)
}
\arguments{
\item{raster_source}{character or list of RasterStack. You can either input a path to a folder with the required rasters or a list of RasterStack organized by year/scenario.}

\item{Perf_args}{named list of strings. Correspondence between PerfFUN arguments and raster names.}

\item{separator}{character. Character that separates variable names, years and scenarios.}

\item{PerfFUN}{function. Function used to modify rasters.}
}
\value{
Returns a list of raster stacks for the variables required, organized by year/scenario combination.
}
\description{
\code{transform_rasters} Transform environmental rasters in ecophysiological rasters.
}
\examples{
FulanusEcoRasters <-
  get_rasters(
    raster_source = "/Users/gabriel/Documents/Mapinguari-development/global_grids_10_minutes",
    ext = FulanusDistribution,
    non_fixed_var = c('prec', 'tmin', 'tmax'),
    fixed_var = 'alt',
    years = c("present", '2050', '2070'),
    scenarios = c('rcp26', 'rcp45', 'rcp85'),
    reorder = TRUE,
    separator = '_')

perf_functions <-
fit_curves(formula = performance ~ s(temp, bs = 'cs') + size,
 data = FulanusPhysiology,
 type = 'GAMM',
 random = list(id = ~ 1),
 separator = '_'
)

Perf_rasters <-
transform_rasters(raster_source = FulanusEcoRasters,
 Perf_args = list(temp = 'tmax', size = 10),
 separator = '_',
 PerfFUN = perf_functions$predict$model_1)

}
