% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_rasters.R
\name{transform_rasters}
\alias{transform_rasters}
\title{Transform environmental rasters in ecophysiological rasters.}
\usage{
transform_rasters(raster_stack, FUN_qlist, separator = "_", alert = NULL)
}
\arguments{
\item{raster_stack}{RasterStack. Stack with environmental layers.}

\item{FUN_qlist}{list. A list of unevaluated expressions, as created by function qlist.}

\item{separator}{character. Character that separates variable names, years and scenarios.}

\item{alert}{integer. Plays a sound alert when function is done running. See documentation of package \code{beepr} for description of sound options.}
}
\value{
Returns a RasterStack with layers for the predictions required.
}
\description{
\code{transform_rasters} Get model predictions for a raster stack
}
\examples{

FulanusEcoRasters_present <-
  get_rasters(
    raster_source = "/Users/gabriel/Documents/Mapinguari/global_grids_10_minutes",
    ext = FulanusDistribution,
    margin = 5,
    non_fixed_var = c('prec', 'tmin', 'tmax'),
    fixed_var = 'alt',
    years = c("present"),
    reorder = TRUE)

transform_rasters(raster_stack = FulanusEcoRasters_present$present,
  FUN = qlist(total_1sem = sum(tmax[[1:6]]),
    mean_1sem = mean(tmax[[1:6]]),
    sd_1sem = sd(tmax[[1:6]]),
    total_2sem = sum(tmax[[7:12]]),
    mean_2sem = mean(tmax[[7:12]]),
    sd_2sem = sd(tmax[[7:12]])),
  separator = '_',
  alert = NULL)

}
